SHAP clustering

SHAP values are used to cluster data samples using the k-means method to identify subgroups of individuals with specific patterns of feature contributions.

Usage

SHAPclust(
  task,
  trained_model,
  splits,
  shap_Mean_wide,
  shap_Mean_long,
  num_of_clusters = 4,
  seed = 246,
  subset = 1,
  algorithm = "Hartigan-Wong",
  iter.max = 1000
)

Arguments

task

an mlr3 task for binary classification

trained_model

an mlr3 trained learner object

splits

an mlr3 object defining data splits for train and test sets

shap_Mean_wide

the data frame of SHAP values in wide format from eSHAP_plot.R

shap_Mean_long

the data frame of SHAP values in long format from eSHAP_plot.R

num_of_clusters

number of clusters to make based on SHAP values, default: 4

seed

an integer for reproducibility, Default to 246

subset

what percentage of the instances to use from 0 to 1 where 1 means all

algorithm

k-means algorithm character: "Hartigan-Wong", "Lloyd", "Forgy", "MacQueen".

iter.max

maximum number of iterations allowed

Value

A list containing four elements:

shap_plot_onerow

An interactive plot displaying the SHAP values for each feature, clustered by the specified number of clusters. Each cluster is shown in a facet.

combined_plot

A ggplot2 figure combining confusion matrices for each cluster, providing insights into the model's performance within each identified subgroup.

kmeans_fvals_desc

A summary table containing statistical descriptions of the clusters based on feature values.

shap_Mean_wide_kmeans

A data frame containing clustered SHAP values along with predictions and ground truth information.

kmeans_info

Information about the k-means clustering process, including cluster centers and assignment details.

References

Zargari Marandi, R., 2024. ExplaineR: an R package to explain machine learning models. Bioinformatics advances, 4(1), p.vbae049, https://doi.org/10.1093/bioadv/vbae049.

See also

Examples

# \donttest{
library("explainer")
seed <- 246
set.seed(seed)
# Load necessary packages
if (!requireNamespace("mlbench", quietly = TRUE)) stop("mlbench not installed.")
if (!requireNamespace("mlr3learners", quietly = TRUE)) stop("mlr3learners not installed.")
if (!requireNamespace("ranger", quietly = TRUE)) stop("ranger not installed.")
# Load BreastCancer dataset
utils::data("BreastCancer", package = "mlbench")
target_col <- "Class"
positive_class <- "malignant"
mydata <- BreastCancer[, -1]
mydata <- na.omit(mydata)
sex <- sample(
  c("Male", "Female"),
  size = nrow(mydata),
  replace = TRUE
)
mydata$age <- as.numeric(sample(
  seq(18, 60),
  size = nrow(mydata),
  replace = TRUE
))
mydata$sex <- factor(
  sex,
  levels = c("Male", "Female"),
  labels = c(1, 0)
)
maintask <- mlr3::TaskClassif$new(
  id = "my_classification_task",
  backend = mydata,
  target = target_col,
  positive = positive_class
)
splits <- mlr3::partition(maintask)
mylrn <- mlr3::lrn(
  "classif.ranger",
  predict_type = "prob"
)
mylrn$train(maintask, splits$train)
SHAP_output <- eSHAP_plot(
  task = maintask,
  trained_model = mylrn,
  splits = splits,
  sample.size = 2, # also 30 or more
  seed = seed,
  subset = 0.02 # up to 1
)
#> Warning: Ignoring unknown aesthetics: text
shap_Mean_wide <- SHAP_output[[2]]
shap_Mean_long <- SHAP_output[[3]]
SHAP_plot_clusters <- SHAPclust(
  task = maintask,
  trained_model = mylrn,
  splits = splits,
  shap_Mean_wide = shap_Mean_wide,
  shap_Mean_long = shap_Mean_long,
  num_of_clusters = 3, # your choice
  seed = seed,
  subset = 0.02, # match with eSHAP_plot
  algorithm = "Hartigan-Wong",
  iter.max = 10
)
#> Key: <sample_num, feature, Phi>
#>     sample_num         feature           Phi cluster    mean_phi     f_val
#>          <int>          <char>         <num>   <int>       <num>     <num>
#>  1:          1     Bare.nuclei -0.0502892857       1 0.131418095 0.2500000
#>  2:          1     Bl.cromatin  0.0067321429       1 0.099304206 0.4285714
#>  3:          1      Cell.shape  0.1119261905       1 0.103042937 1.0000000
#>  4:          1       Cell.size  0.3119706349       1 0.194636190 1.0000000
#>  5:          1    Cl.thickness  0.0715214286       1 0.061621905 0.8750000
#>  6:          1    Epith.c.size  0.0469928571       1 0.042565079 1.0000000
#>  7:          1   Marg.adhesion  0.1016091270       1 0.084125397 0.5714286
#>  8:          1         Mitoses  0.0398400794       1 0.007968016 1.0000000
#>  9:          1 Normal.nucleoli  0.0343305556       1 0.039085476 1.0000000
#> 10:          1             age  0.0000000000       1 0.008259048 0.0000000
#> 11:          1             sex  0.0036666667       1 0.002500317 0.0000000
#> 12:          2     Bare.nuclei -0.1375361111       3 0.131418095 0.0000000
#> 13:          2     Bl.cromatin -0.1114793651       3 0.099304206 0.0000000
#> 14:          2      Cell.shape -0.0514559524       3 0.103042937 0.1428571
#> 15:          2       Cell.size -0.0857575397       3 0.194636190 0.0000000
#> 16:          2    Cl.thickness  0.0169722222       3 0.061621905 0.3750000
#> 17:          2    Epith.c.size -0.0024634921       3 0.042565079 0.0000000
#> 18:          2   Marg.adhesion -0.0533325397       3 0.084125397 0.0000000
#> 19:          2         Mitoses  0.0000000000       3 0.007968016 0.0000000
#> 20:          2 Normal.nucleoli -0.0223630952       3 0.039085476 0.0000000
#> 21:          2             age -0.0156880952       3 0.008259048 0.8205128
#> 22:          2             sex  0.0000000000       3 0.002500317 1.0000000
#> 23:          3     Bare.nuclei  0.1458841270       2 0.131418095 1.0000000
#> 24:          3     Bl.cromatin  0.1523980159       2 0.099304206 0.8571429
#> 25:          3      Cell.shape  0.1557253968       2 0.103042937 0.8571429
#> 26:          3       Cell.size  0.2166472222       2 0.194636190 0.5714286
#> 27:          3    Cl.thickness  0.0117880952       2 0.061621905 0.3750000
#> 28:          3    Epith.c.size  0.1036420635       2 0.042565079 1.0000000
#> 29:          3   Marg.adhesion  0.1224638889       2 0.084125397 1.0000000
#> 30:          3         Mitoses  0.0000000000       2 0.007968016 0.0000000
#> 31:          3 Normal.nucleoli  0.0616575397       2 0.039085476 0.3333333
#> 32:          3             age  0.0057753968       2 0.008259048 1.0000000
#> 33:          3             sex  0.0000000000       2 0.002500317 1.0000000
#> 34:          4     Bare.nuclei  0.0890119048       2 0.131418095 1.0000000
#> 35:          4     Bl.cromatin  0.0967761905       2 0.099304206 1.0000000
#> 36:          4      Cell.shape  0.0786674603       2 0.103042937 0.5714286
#> 37:          4       Cell.size  0.1656662698       2 0.194636190 0.7142857
#> 38:          4    Cl.thickness  0.1231694444       2 0.061621905 1.0000000
#> 39:          4    Epith.c.size  0.0395043651       2 0.042565079 0.7500000
#> 40:          4   Marg.adhesion  0.0526595238       2 0.084125397 1.0000000
#> 41:          4         Mitoses  0.0000000000       2 0.007968016 0.0000000
#> 42:          4 Normal.nucleoli  0.0107317460       2 0.039085476 0.5555556
#> 43:          4             age  0.0001428571       2 0.008259048 0.1794872
#> 44:          4             sex  0.0000000000       2 0.002500317 1.0000000
#> 45:          5     Bare.nuclei -0.2343690476       3 0.131418095 0.0000000
#> 46:          5     Bl.cromatin -0.1291353175       3 0.099304206 0.0000000
#> 47:          5      Cell.shape -0.1174396825       3 0.103042937 0.0000000
#> 48:          5       Cell.size -0.1931392857       3 0.194636190 0.0000000
#> 49:          5    Cl.thickness -0.0846583333       3 0.061621905 0.0000000
#> 50:          5    Epith.c.size -0.0202226190       3 0.042565079 0.0000000
#> 51:          5   Marg.adhesion -0.0905619048       3 0.084125397 0.0000000
#> 52:          5         Mitoses  0.0000000000       3 0.007968016 0.0000000
#> 53:          5 Normal.nucleoli -0.0663444444       3 0.039085476 0.0000000
#> 54:          5             age -0.0196888889       3 0.008259048 0.7179487
#> 55:          5             sex  0.0088349206       3 0.002500317 0.0000000
#>     sample_num         feature           Phi cluster    mean_phi     f_val
#>     unscaled_f_val correct_prediction pred_prob pred_class
#>              <num>             <fctr>     <num>     <fctr>
#>  1:              2            Correct 0.9646984  malignant
#>  2:              4            Correct 0.9646984  malignant
#>  3:              8            Correct 0.9646984  malignant
#>  4:              8            Correct 0.9646984  malignant
#>  5:              9            Correct 0.9646984  malignant
#>  6:              6            Correct 0.9646984  malignant
#>  7:              5            Correct 0.9646984  malignant
#>  8:              4            Correct 0.9646984  malignant
#>  9:             10            Correct 0.9646984  malignant
#> 10:             19            Correct 0.9646984  malignant
#> 11:              1            Correct 0.9646984  malignant
#> 12:              1            Correct 0.0000000     benign
#> 13:              1            Correct 0.0000000     benign
#> 14:              2            Correct 0.0000000     benign
#> 15:              1            Correct 0.0000000     benign
#> 16:              5            Correct 0.0000000     benign
#> 17:              2            Correct 0.0000000     benign
#> 18:              1            Correct 0.0000000     benign
#> 19:              1            Correct 0.0000000     benign
#> 20:              1            Correct 0.0000000     benign
#> 21:             51            Correct 0.0000000     benign
#> 22:              2            Correct 0.0000000     benign
#> 23:             10            Correct 0.9759817  malignant
#> 24:              7            Correct 0.9759817  malignant
#> 25:              7            Correct 0.9759817  malignant
#> 26:              5            Correct 0.9759817  malignant
#> 27:              5            Correct 0.9759817  malignant
#> 28:              6            Correct 0.9759817  malignant
#> 29:              8            Correct 0.9759817  malignant
#> 30:              1            Correct 0.9759817  malignant
#> 31:              4            Correct 0.9759817  malignant
#> 32:             58            Correct 0.9759817  malignant
#> 33:              2            Correct 0.9759817  malignant
#> 34:             10            Correct 1.0000000  malignant
#> 35:              8            Correct 1.0000000  malignant
#> 36:              5            Correct 1.0000000  malignant
#> 37:              6            Correct 1.0000000  malignant
#> 38:             10            Correct 1.0000000  malignant
#> 39:              5            Correct 1.0000000  malignant
#> 40:              8            Correct 1.0000000  malignant
#> 41:              1            Correct 1.0000000  malignant
#> 42:              6            Correct 1.0000000  malignant
#> 43:             26            Correct 1.0000000  malignant
#> 44:              2            Correct 1.0000000  malignant
#> 45:              1            Correct 0.0010000     benign
#> 46:              1            Correct 0.0010000     benign
#> 47:              1            Correct 0.0010000     benign
#> 48:              1            Correct 0.0010000     benign
#> 49:              2            Correct 0.0010000     benign
#> 50:              2            Correct 0.0010000     benign
#> 51:              1            Correct 0.0010000     benign
#> 52:              1            Correct 0.0010000     benign
#> 53:              1            Correct 0.0010000     benign
#> 54:             47            Correct 0.0010000     benign
#> 55:              1            Correct 0.0010000     benign
#>     unscaled_f_val correct_prediction pred_prob pred_class
#> Warning: Ignoring unknown aesthetics: text
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: Groups with fewer than two datapoints have been dropped.
#> ℹ Set `drop = FALSE` to consider such groups for position adjustment purposes.
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: Computation failed in `stat_ydensity()`.
#> Caused by error in `$<-.data.frame`:
#> ! replacement has 1 row, data has 0
#> Warning: 'ggimage' is missing. Will not plot arrows and zero-shading.
#> Warning: 'rsvg' is missing. Will not plot arrows and zero-shading.
#> Warning: 'ggimage' is missing. Will not plot arrows and zero-shading.
#> Warning: 'rsvg' is missing. Will not plot arrows and zero-shading.
#> Warning: 'ggimage' is missing. Will not plot arrows and zero-shading.
#> Warning: 'rsvg' is missing. Will not plot arrows and zero-shading.
# }